--- docs/source/io.rst
@@ -1,20 +1,0 @@
-.. _IO:
-
-I/O
----
-
-The I/O is handled by the :ref:`io_handler <iomodule>` module.
-
-.. _iomodule:
-
-io_handler
-**********
-
-.. automodule:: io.io_handler
-   :members:
-
-io_hdf5
-*******
-
-.. automodule:: io.io_hdf5
-   :members:
