+++ simtools/applications/compare_cumulative_psf.py
@@ -87,7 +87,7 @@ import yaml
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.model.telescope_model import TelescopeModel
 from simtools.ray_tracing import RayTracing
 from simtools.visualization import visualize
