+++ simtools/applications/generate_corsika_histograms.py
@@ -173,7 +173,7 @@ import simtools.utils.general as gen
 from simtools.configuration import configurator
 from simtools.corsika import corsika_histograms_visualize
 from simtools.corsika.corsika_histograms import CorsikaHistograms
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 
 logger = logging.getLogger()
