+++ simtools/applications/generate_simtel_array_histograms.py
@@ -56,7 +56,7 @@ from matplotlib.backends.backend_pdf import PdfPages
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.simtel.simtel_histograms import SimtelHistograms
