+++ simtools/applications/make_regular_arrays.py
@@ -42,7 +42,7 @@ import astropy.units as u
 import simtools.utils.general as gen
 from simtools import db_handler
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.layout.layout_array import LayoutArray
