+++ simtools/applications/plot_layout_array.py
@@ -45,7 +45,7 @@ from astropy import units as u
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.layout.layout_array import LayoutArray
 from simtools.visualization.visualize import plot_array
