+++ simtools/applications/tune_psf.py
@@ -95,7 +95,7 @@ from matplotlib.backends.backend_pdf import PdfPages
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.model.model_utils import split_simtel_parameter
 from simtools.model.telescope_model import TelescopeModel
 from simtools.ray_tracing import RayTracing
