+++ simtools/applications/validate_camera_fov.py
@@ -60,7 +60,7 @@ from pathlib import Path
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.model.telescope_model import TelescopeModel
