+++ simtools/applications/validate_optics.py
@@ -78,7 +78,7 @@ from matplotlib.backends.backend_pdf import PdfPages
 
 import simtools.utils.general as gen
 from simtools.configuration import configurator
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.model.telescope_model import TelescopeModel
 from simtools.ray_tracing import RayTracing
