+++ simtools/model/array_model.py
@@ -2,7 +2,7 @@ import logging
 from copy import copy
 
 from simtools import db_handler
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.layout.layout_array import LayoutArray
 from simtools.model.telescope_model import TelescopeModel
 from simtools.simtel.simtel_config_writer import SimtelConfigWriter
