+++ simtools/model/telescope_model.py
@@ -10,7 +10,7 @@ from astropy.table import Table
 
 import simtools.utils.general as gen
 from simtools import db_handler
-from simtools.io import io_handler
+from simtools.io_operations import io_handler
 from simtools.model.camera import Camera
 from simtools.model.mirrors import Mirrors
 from simtools.simtel.simtel_config_writer import SimtelConfigWriter
